first iteration Objective system
made ObjectiveScriptableObject script which allows for the creation of scriptable objects that derive from it in the editor by going clicking on create, going to ScriptableObjects option, then choosing the "objective" option. it has 4 public variables and also creates a UnityEvent<int> variable which can be used to alert other scripts when the objective has been completed. I also made the GoalPost script which should be attached to gameobjects that the player is supposed to reach. Make sure player has the tag "Player" so that it works properly. The script needs to be given a reference to ObjectiveScriptableObject so that the UnityEvent can be called.