It works It works IT WORKS IT WORKS!!!!!!!!! MY PAIN IS OVER (almost) IT FINALLY WORKS
fix a bug in the vector math to define the wrong plane for measuring the roll
angle.

Given that this is the end of the quarter I will like have this tested now,
though the code is way to messing for the branch to be merge ready. I need
to clean up and document my code before making a pull request.