updated save system
new save system uses the eventManager to handle the event of a player's death. got rid of savePoint script and save SO. data for the player to respawn is inside of the Player script and handled there as well. NarrrationManager's narrationPaused bool is set to true when a player dies so the narration won't get stuck. I still need to make it so that an onfinished objective is reset if death occurs mid objective