Finished implementing alternate hover, used an exponent function of time to determine the fall rate. You may set the base and the coefficient of time in the inspector.
Attempted to added a factor coefficient to reduce sliding
after move button is released. Though it appears to work I did not do it
correctly. Will return to it in a later commit.

Finally, I added a particle system to Ika's prefab to visualize the wind. It
create white lines only when flying that go in the direction of relative
wind. Currently the speed and number of these white lines are constant. I
plan on making it depend on Ika's velocity.