fully functional Melee attack implemented
The melee attack system is ready for testing. the OldManMcgee behavior tree scriptable object has the behavior I was messing with. You can create an attackNode and set it's mask so that the AI only attacks certain enteties. The Melee class, which inherits from the Spell class, has a cooldown variable that can be change to change the cooldown for the atatck. Create Instant damage scriptable objects and add them to the list of spell_effects in the Melee class. These will be applied to the Enemy in order.