made AI system independent of scene(Beware of Bug!) (#38)
The node set ups were moved to the Awake method, so there is no need for the narration manager to let notify when the the scene has loaded.  you can test this in the test AI settings. There is a Bug tho. When Ika's health reaches 0, it rapidly continues to drop. This seems to be because Ika's respawn is dependent on the narration. In the oldManMcgee behavior tree, you can change the destination from Ika to tree in the goToNode, and change the mask to tree in the AttackNode. Doing this will make oldManMcgee attack one of the trees, as long as their tag is Tree and their layer is Tree. It works as intended when it attacks the tree, so long as it has a health manager and the box collider is on. But there is unexpected behavior when Ika's life reaches zero.