implemented Projectile spell
The projectile Spell has been impleminted and giving an object the MageEntity script will make the projectile class a required component. It must be assigned a behaviour tree, which can be made in the Behaviour tree editor window. Make a projectile node and set it's target to IKA. This causes the NavMesh agent to first turn to face the target, then shoot a projectile. Remember to also set the projectile GameObject in the Projectile script. I provided one labled Projectile1 in the Prefabs folder. Attached to it is a script called Projectile logic which handles itself.